Fix VAT calculation check to allow for rounding; update import logic

+++ database/migrations/2025_07_02_153830_create_invoice_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('invoice_id')->constrained('invoices')->onDelete('cascade');
             $table->string('description')->nullable();
             $table->string('item_name', 100);
-            $table->integer('quantity');
+            $table->decimal('quantity', 12, 2);
             $table->decimal('price', 12, 2);
             $table->decimal('total', 12, 2);
             $table->string('unit')->nullable();
